Introduces Lazy as a concept that can be applied to any route
Once set, some semantics will change to allow better cooperation between
the server-side and client-side with regards to AJAX "Lazy" calls.

The only implemented semantic change is that error messages are given in
a partial view, allowing the "Lazy" initiator client-side to simply use
the error as given, instead of printing a generic error message.

This builds upon #619.